connect prev recording actions to presets
